move `Helper` class to `dff/utils/testing/telegram`
