+++ tests/messengers/telegram/test_examples.py
@@ -31,10 +31,10 @@ dot_path_to_addon = get_path_from_tests_to_current_dir(__file__, separator=".")
         "7_polling_setup",
     ],
 )
-async def test_client_examples(example_module_name, bot_id, tg_client, helper, tmp_path):
+async def test_client_examples(example_module_name, bot_id, tg_client, testing_helper, tmp_path):
     example_module = importlib.import_module(f"examples.{dot_path_to_addon}.{example_module_name}")
     pipeline = example_module.pipeline
     happy_path = example_module.happy_path
-    test_helper = helper(tg_client, pipeline, bot_id)
+    test_helper = testing_helper(tg_client, pipeline, bot_id)
     logging.info("Test start")
     await test_helper.check_happy_path(happy_path, tmp_path)
